Fix share Intent not closing (#633)
On a couple of devices I have tested the sample application and I noticed that when I pressed share in the Tracks section and tried to exit the share Intent would launch multiple times and enter a loop until the system realises this. Returning true here tells the system that the menu item press was handled correctly.